+++ NEWS.md
@@ -2,12 +2,14 @@
 
 ## New features
 
+- `eta_squared()` family now indicate the type of sum-of-squares used.
 - `rank_biserial()` estimates CIs using the normal approximation (previously used bootstrapping).  
 - `hedges_g()` now used exact bias correction (thanks to @mdelacre for the suggestion!)  
 - `glass_delta()` now estimates CIs using the NCP method based on Algina et al (2006).
 
 ## Bug fixes
 
+- `eta_squared()` family returns correctly returns the type 2/3 effect sizes for mixed ANOVAs fit with `afex`.
 - `cohens_d()` family now correctly deals with missing factor levels ( #318 )
 - `cohens_d()` / `hedges_g()` minor fix for CI with unequal variances.